refactor(brutus): extract credential research to research.go
Move 3 research functions (HTTP auth, browser, LLM).
Part of main.go refactoring (5/9).

Related: C1/C2 (split 1340-line main.go)
