+++ cmd/brutus/main.go
@@ -16,13 +16,10 @@ package main
 
 import (
 	"bufio"
-	"context"
-	"crypto/tls"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"io"
-	"net/http"
 	"os"
 	"strings"
 	"time"
@@ -31,9 +28,7 @@ import (
 
 	// Import plugins and analyzers to register them
 	_ "github.com/praetorian-inc/brutus/internal/analyzers"
-	"github.com/praetorian-inc/brutus/internal/analyzers/vision"
 	_ "github.com/praetorian-inc/brutus/internal/plugins"
-	"github.com/praetorian-inc/brutus/internal/plugins/browser"
 	"github.com/praetorian-inc/brutus/internal/plugins/snmp"
 )
 
@@ -549,254 +544,6 @@ func runSingleTarget(target, protocol string, base *baseConfigOptions) ([]brutus
 	}
 
 	return results, hasSuccess
-}
-
-// detectHTTPAuthTypeWithBanner probes an HTTP target to determine the authentication type.
-// Returns auth type ("basic", "form", or "" if not HTTP) and the banner text for LLM analysis.
-func detectHTTPAuthTypeWithBanner(target string, useHTTPS bool, timeout time.Duration, verbose bool) (authType, banner string) {
-	scheme := "http"
-	if useHTTPS {
-		scheme = "https"
-	}
-	url := fmt.Sprintf("%s://%s/", scheme, target)
-
-	client := &http.Client{
-		Timeout: timeout,
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-		},
-		CheckRedirect: func(req *http.Request, via []*http.Request) error {
-			return http.ErrUseLastResponse
-		},
-	}
-
-	req, err := http.NewRequest("GET", url, http.NoBody)
-	if err != nil {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] HTTP probe error creating request: %v\n", err)
-		}
-		return "", "" // Not HTTP or error
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] HTTP probe error: %v\n", err)
-		}
-		return "", "" // Not HTTP or error
-	}
-	defer resp.Body.Close()
-
-	// Build banner from response headers and body
-	var bannerBuilder strings.Builder
-	bannerBuilder.WriteString(fmt.Sprintf("HTTP/%d.%d %s\n", resp.ProtoMajor, resp.ProtoMinor, resp.Status))
-
-	// Include relevant headers
-	for _, header := range []string{"Server", "WWW-Authenticate", "X-Powered-By", "X-Server", "X-AspNet-Version"} {
-		if val := resp.Header.Get(header); val != "" {
-			bannerBuilder.WriteString(fmt.Sprintf("%s: %s\n", header, val))
-		}
-	}
-
-	// Read body (limited to avoid memory issues)
-	body := make([]byte, 4096)
-	n, _ := io.ReadFull(resp.Body, body)
-	if n > 0 {
-		bannerBuilder.WriteString("\n")
-		bannerBuilder.Write(body[:n])
-	}
-
-	banner = bannerBuilder.String()
-
-	// Check for WWW-Authenticate header (indicates Basic Auth)
-	if authHeader := resp.Header.Get("WWW-Authenticate"); authHeader != "" {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] HTTP probe: WWW-Authenticate header found: %s\n", authHeader)
-		}
-		return "basic", banner
-	}
-
-	// Check for 401 status without WWW-Authenticate (some servers don't send it)
-	if resp.StatusCode == http.StatusUnauthorized {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] HTTP probe: 401 without WWW-Authenticate, assuming basic\n")
-		}
-		return "basic", banner
-	}
-
-	return "form", banner
-}
-
-// researchBrowserCredentials uses Claude Vision + Perplexity for browser-based credential research.
-// 1. Claude Vision analyzes the screenshot to identify the application
-// 2. Perplexity researches default credentials for the identified application
-// Returns the researched credentials AND the configured browser plugin (for use in Brute).
-func researchBrowserCredentials(target string, base *baseConfigOptions) ([]brutus.Credential, *browser.Plugin) {
-	if base.llmConfig == nil || !base.llmConfig.Enabled {
-		return nil, nil
-	}
-
-	// Get API keys
-	perplexityKey := os.Getenv("PERPLEXITY_API_KEY")
-	anthropicKey := os.Getenv("ANTHROPIC_API_KEY")
-
-	// Create the browser plugin with configured analyzers
-	browserPlugin := &browser.Plugin{
-		Timeout:         60 * time.Second,
-		TabCount:        base.browserTabs,
-		PageLoadTimeout: 15 * time.Second,
-		UseHTTPS:        base.useHTTPS,
-		Visible:         base.browserVisible,
-		Verbose:         base.verbose,
-		AIVerify:        base.aiVerify,
-	}
-
-	// Configure Vision analyzer (Claude) for screenshot analysis
-	if anthropicKey != "" {
-		browserPlugin.VisionAnalyzer = &vision.Client{
-			APIKey: anthropicKey,
-		}
-		if base.verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] Configured Claude Vision for screenshot analysis\n")
-		}
-	} else if base.verbose {
-		fmt.Fprintf(os.Stderr, "[verbose] No ANTHROPIC_API_KEY - skipping vision analysis\n")
-	}
-
-	// Configure Credential researcher (Perplexity) for default credential lookup
-	if perplexityKey != "" {
-		factory := brutus.GetAnalyzerFactory("perplexity")
-		if factory != nil {
-			analyzer := factory(&brutus.LLMConfig{
-				Enabled:  true,
-				Provider: "perplexity",
-				APIKey:   perplexityKey,
-			})
-			if credAnalyzer, ok := analyzer.(brutus.CredentialAnalyzer); ok {
-				browserPlugin.CredentialResearcher = credAnalyzer
-				if base.verbose {
-					fmt.Fprintf(os.Stderr, "[verbose] Configured Perplexity for credential research\n")
-				}
-			}
-		}
-	} else if base.verbose {
-		fmt.Fprintf(os.Stderr, "[verbose] No PERPLEXITY_API_KEY - skipping credential research\n")
-	}
-
-	// Call AnalyzePage to do the full two-stage AI flow
-	ctx, cancel := context.WithTimeout(context.Background(), 90*time.Second)
-	defer cancel()
-
-	if base.verbose {
-		fmt.Fprintf(os.Stderr, "[verbose] Starting headless browser (this may take a few seconds on first run)...\n")
-	}
-
-	pageAnalysis, credentials, err := browserPlugin.AnalyzePage(ctx, target)
-	if err != nil {
-		if base.verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] Page analysis error: %v\n", err)
-		}
-		return nil, nil
-	}
-
-	// Log analysis results
-	if base.verbose && pageAnalysis != nil {
-		fmt.Fprintf(os.Stderr, "[verbose] Vision detected: %s %s %s (login page: %v, confidence: %.2f)\n",
-			pageAnalysis.Application.Vendor,
-			pageAnalysis.Application.Model,
-			pageAnalysis.Application.Type,
-			pageAnalysis.IsLoginPage,
-			pageAnalysis.Confidence)
-	}
-
-	if len(credentials) > 0 && base.verbose {
-		fmt.Fprintf(os.Stderr, "[verbose] AI researched %d credential pairs:\n", len(credentials))
-		for _, c := range credentials {
-			fmt.Fprintf(os.Stderr, "[verbose]   %s:%s\n", c.Username, c.Password)
-		}
-	}
-
-	return credentials, browserPlugin
-}
-
-// researchCredentialsWithLLM uses the configured LLM to research default credentials for a target
-func researchCredentialsWithLLM(target, banner string, llmConfig *brutus.LLMConfig, verbose bool) []brutus.Credential {
-	if llmConfig == nil || !llmConfig.Enabled {
-		return nil
-	}
-
-	// Get the analyzer factory for the configured provider
-	factory := brutus.GetAnalyzerFactory(llmConfig.Provider)
-	if factory == nil {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] LLM analyzer not found for provider: %s\n", llmConfig.Provider)
-		}
-		return nil
-	}
-	analyzer := factory(llmConfig)
-
-	// Check if analyzer supports credential pairs (CredentialAnalyzer interface)
-	credAnalyzer, ok := analyzer.(brutus.CredentialAnalyzer)
-	if !ok {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] LLM analyzer doesn't support credential pairs, using password-only\n")
-		}
-		// Fall back to password-only analysis
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-
-		bannerInfo := brutus.BannerInfo{
-			Protocol: "http",
-			Target:   target,
-			Banner:   banner,
-		}
-
-		passwords, err := analyzer.Analyze(ctx, bannerInfo)
-		if err != nil {
-			if verbose {
-				fmt.Fprintf(os.Stderr, "[verbose] LLM analysis error: %v\n", err)
-			}
-			return nil
-		}
-
-		// Convert passwords to credentials with common usernames
-		var creds []brutus.Credential
-		for _, pwd := range passwords {
-			creds = append(creds,
-				brutus.Credential{Username: "admin", Password: pwd},
-				brutus.Credential{Username: "root", Password: pwd})
-		}
-		return creds
-	}
-
-	// Use CredentialAnalyzer for full username:password pairs
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	bannerInfo := brutus.BannerInfo{
-		Protocol: "http",
-		Target:   target,
-		Banner:   banner,
-	}
-
-	creds, err := credAnalyzer.AnalyzeCredentials(ctx, bannerInfo)
-	if err != nil {
-		if verbose {
-			fmt.Fprintf(os.Stderr, "[verbose] LLM credential analysis error: %v\n", err)
-		}
-		return nil
-	}
-
-	if verbose && len(creds) > 0 {
-		fmt.Fprintf(os.Stderr, "[verbose] LLM suggested credentials:\n")
-		for _, c := range creds {
-			fmt.Fprintf(os.Stderr, "[verbose]   %s:%s\n", c.Username, c.Password)
-		}
-	}
-
-	return creds
 }
 
 func loadPasswords(inline, file string, inlineFlagSet bool) []string {
